scripts now require repo uri

+++ delete-datasets.sh
@@ -1,19 +1,27 @@
-mvn cdk:delete-dataset \
-  -Dcdk.repositoryUri=repo:hbase:localhost.localdomain \
-  -Dcdk.datasetName=party \
-  -Dcdk.avroSchemaFile=src/main/avro/party.avsc
+#!/bin/bash
+
+if [ $# -ne 1 ]; then
+  echo "usage: $0 <hbase repositoryUri>"
+  echo "e.g.:"
+  echo "  $0 repo:hbase:zk"
+  echo "  $0 repo:hbase:zk1,zk2,zk3"  
+  exit 1
+fi
+
+echo "using repositoryUri: $1"
 
 mvn cdk:delete-dataset \
-  -Dcdk.repositoryUri=repo:hbase:localhost.localdomain \
+  -Dcdk.repositoryUri=$1 \
+  -Dcdk.datasetName=party
+
+mvn cdk:delete-dataset \
+  -Dcdk.repositoryUri=$1 \
   -Dcdk.datasetName=event \
-  -Dcdk.avroSchemaFile=src/main/avro/event.avsc
 
 mvn cdk:delete-dataset \
-  -Dcdk.repositoryUri=repo:hbase:localhost.localdomain \
+  -Dcdk.repositoryUri=$1 \
   -Dcdk.datasetName=address \
-  -Dcdk.avroSchemaFile=src/main/avro/address.avsc
 
 mvn cdk:delete-dataset \
-  -Dcdk.repositoryUri=repo:hbase:localhost.localdomain \
+  -Dcdk.repositoryUri=$1 \
   -Dcdk.datasetName=party_address \
-  -Dcdk.avroSchemaFile=src/main/avro/party_address.avsc
